requirements.txt
numpy==2.3.3
branca==0.8.1
folium==0.20.0
geopandas==1.1.1
leafmap==0.52.2
mapclassify==2.10.0
streamlit-folium==0.25.2
rasterio==1.4.3
localtileserver
setuptools
psutil
rioxarray